+++ database/migrations/2022_03_25_111323_create_comments_table.php
@@ -19,8 +19,8 @@ class CreateCommentsTable extends Migration
             $table->string('title');
             $table->text('content');
             $table->integer('rating');
-            $table->integer('user_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->timestamps();
             $table->softDeletes();
             $table->foreign('user_id')->references('id')->on('users');
